Challenge_1b/requirements.txt
# requirements.txt

# Core ML Libraries
torch
transformers
sentence-transformers
rank_bm25
gliner-spacy

# PDF Processing
PyMuPDF
pdfplumber

# Utility
jsonschema